Rename models for surah and detail surah, remove unnecessary modules and files
