--- lib/app/modules/detail_book/bindings/detail_book_binding.dart
@@ -1,12 +1,0 @@
-import 'package:get/get.dart';
-
-import '../controllers/detail_book_controller.dart';
-
-class DetailBookBinding extends Bindings {
-  @override
-  void dependencies() {
-    Get.lazyPut<DetailBookController>(
-      () => DetailBookController(),
-    );
-  }
-}
