+++ lib/app/modules/detail_juz/controllers/detail_juz_controller.dart
@@ -4,7 +4,7 @@ import 'package:just_audio/just_audio.dart';
 import 'package:quran/app/components/snackbar.dart';
 import 'package:quran/app/data/db/bookmark.dart';
 import 'package:quran/app/data/models/juz.dart';
-import 'package:quran/app/data/models/surah_main.dart';
+import 'package:quran/app/data/models/surah.dart';
 import 'package:sqflite/sqlite_api.dart';
 
 class DetailJuzController extends GetxController {
