--- lib/app/modules/search/bindings/search_binding.dart
@@ -1,12 +1,0 @@
-import 'package:get/get.dart';
-
-import '../controllers/search_controller.dart';
-
-class SearchBinding extends Bindings {
-  @override
-  void dependencies() {
-    Get.lazyPut<SearchController>(
-      () => SearchController(),
-    );
-  }
-}
