--- lib/app/modules/search/controllers/search_controller.dart
@@ -1,3 +1,0 @@
-import 'package:get/get.dart';
-
-class SearchController extends GetxController {}
