+++ lib/app/routes/app_pages.dart
@@ -1,7 +1,5 @@
 import 'package:get/get.dart';
 
-import '../modules/detail_book/bindings/detail_book_binding.dart';
-import '../modules/detail_book/views/detail_book_view.dart';
 import '../modules/detail_juz/bindings/detail_juz_binding.dart';
 import '../modules/detail_juz/views/detail_juz_view.dart';
 import '../modules/detail_surah/bindings/detail_surah_binding.dart';
@@ -10,8 +8,6 @@ import '../modules/home/bindings/home_binding.dart';
 import '../modules/home/views/home_view.dart';
 import '../modules/introduction/bindings/introduction_binding.dart';
 import '../modules/introduction/views/introduction_view.dart';
-import '../modules/search/bindings/search_binding.dart';
-import '../modules/search/views/search_view.dart';
 
 part 'app_routes.dart';
 
@@ -32,19 +28,9 @@ class AppPages {
       binding: IntroductionBinding(),
     ),
     GetPage(
-      name: _Paths.DETAIL_BOOK,
-      page: () => DetailBookView(),
-      binding: DetailBookBinding(),
-    ),
-    GetPage(
       name: _Paths.DETAIL_SURAH,
       page: () => DetailSurahView(),
       binding: DetailSurahBinding(),
-    ),
-    GetPage(
-      name: _Paths.SEARCH,
-      page: () => SearchView(),
-      binding: SearchBinding(),
     ),
     GetPage(
       name: _Paths.DETAIL_JUZ,
