+++ lib/app/routes/app_routes.dart
@@ -5,9 +5,7 @@ abstract class Routes {
   Routes._();
   static const HOME = _Paths.HOME;
   static const INTRODUCTION = _Paths.INTRODUCTION;
-  static const DETAIL_BOOK = _Paths.DETAIL_BOOK;
   static const DETAIL_SURAH = _Paths.DETAIL_SURAH;
-  static const SEARCH = _Paths.SEARCH;
   static const DETAIL_JUZ = _Paths.DETAIL_JUZ;
 }
 
@@ -15,8 +13,6 @@ abstract class _Paths {
   _Paths._();
   static const HOME = '/home';
   static const INTRODUCTION = '/introduction';
-  static const DETAIL_BOOK = '/detail-book';
   static const DETAIL_SURAH = '/detail-surah';
-  static const SEARCH = '/search';
   static const DETAIL_JUZ = '/detail-juz';
 }
